Use valid library folder name
The Arduino IDE does not permit library folders to have spaces. A library under a folder with an invalid name will not be recognized by the Arduino IDE and the IDE will display a warning on startup and whenever Library or Boards Manager is opened or closed:

Ignoring bad library name
The library "SDA Arduino" cannot be used.